#90: refactored CommandLineUsage.
