+++ src/test/java/com/github/romankh3/image/comparison/BaseTest.java
@@ -1,0 +1,29 @@
+package com.github.romankh3.image.comparison;
+
+import static org.junit.Assert.fail;
+
+import java.awt.image.BufferedImage;
+
+/**
+ * Base test for all the tests. Contains common methods, fields for almost all the tests.
+ */
+class BaseTest {
+
+    public void assertImagesEqual(BufferedImage imgA, BufferedImage imgB) {
+        if (imgA.getWidth() != imgB.getWidth() || imgA.getHeight() != imgB.getHeight()) {
+            fail("Images have different dimensions");
+        }
+
+        int width = imgA.getWidth();
+        int height = imgA.getHeight();
+
+        for (int y = 0; y < height; y++) {
+            for (int x = 0; x < width; x++) {
+                if (imgA.getRGB(x, y) != imgB.getRGB(x, y)) {
+                    fail("Images are different, found different pixel at: x = " + x + ", y = " + y);
+                }
+            }
+        }
+    }
+
+}
