+++ src/test/java/com/github/romankh3/image/comparison/ImageComparisonUnitTest.java
@@ -5,24 +5,18 @@ import static com.github.romankh3.image.comparison.model.ComparisonState.MATCH;
 import static com.github.romankh3.image.comparison.model.ComparisonState.MISSMATCH;
 import static com.github.romankh3.image.comparison.model.ComparisonState.SIZE_MISSMATCH;
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
 
 import com.github.romankh3.image.comparison.model.ComparisonResult;
 import java.awt.image.BufferedImage;
-import java.io.File;
 import java.io.IOException;
 import java.net.URISyntaxException;
-import java.nio.file.Files;
-import java.util.concurrent.atomic.AtomicBoolean;
 import org.junit.Test;
 
 /**
  * Unit-level testing for {@link ImageComparison} object.
  */
-public class ImageComparisonUnitTest {
+public class ImageComparisonUnitTest extends BaseTest {
 
     /**
      * The most important test. Shown, that the changes in algorithm,
@@ -175,88 +169,6 @@ public class ImageComparisonUnitTest {
     }
 
     @Test
-    public void testCreateDefault() throws IOException, URISyntaxException {
-        ImageComparison comparison = CommandLineUtil.createDefault();
-        assertImagesEqual(readImageFromResources("image1.png"), comparison.getImage1());
-        assertImagesEqual(readImageFromResources("image2.png"), comparison.getImage2());
-        assertFalse(comparison.getDestination().isPresent());
-    }
-
-    @Test
-    public void testCreateWithTwoArgs() throws IOException, URISyntaxException {
-        File image1 = new File(ImageComparison.class.getClassLoader().getResource("image1.png").toURI().getPath());
-        File image2 = new File(ImageComparison.class.getClassLoader().getResource("image2.png").toURI().getPath());
-        ImageComparison comparison = CommandLineUtil.create(image1.getAbsolutePath(), image2.getAbsolutePath());
-
-        assertImagesEqual(readImageFromResources("image1.png"), comparison.getImage1());
-        assertImagesEqual(readImageFromResources("image2.png"), comparison.getImage2());
-        assertFalse(comparison.getDestination().isPresent());
-    }
-
-    @Test
-    public void testCreateWithTwoImagesAsArgs() throws IOException, URISyntaxException {
-        File image1 = new File(ImageComparison.class.getClassLoader().getResource("image1.png").toURI().getPath());
-        File image2 = new File(ImageComparison.class.getClassLoader().getResource("image2.png").toURI().getPath());
-        ImageComparison comparison = CommandLineUtil.create(new ArgsParser.Arguments(image1, image2, null));
-
-        assertImagesEqual(readImageFromResources("image1.png"), comparison.getImage1());
-        assertImagesEqual(readImageFromResources("image2.png"), comparison.getImage2());
-        assertFalse(comparison.getDestination().isPresent());
-    }
-
-    @Test
-    public void testCreateWithTwoImagesAndDestinationFileAsArgs() throws IOException, URISyntaxException {
-        File image1 = new File(ImageComparison.class.getClassLoader().getResource("image1.png").toURI().getPath());
-        File image2 = new File(ImageComparison.class.getClassLoader().getResource("image2.png").toURI().getPath());
-        File destination = Files.createTempFile("image-comparison-test", ".png").toFile();
-        ImageComparison comparison = CommandLineUtil.create(new ArgsParser.Arguments(image1, image2, destination));
-
-        assertImagesEqual(readImageFromResources("image1.png"), comparison.getImage1());
-        assertImagesEqual(readImageFromResources("image2.png"), comparison.getImage2());
-        assertTrue(comparison.getDestination().isPresent());
-        assertEquals(destination, comparison.getDestination().get());
-    }
-
-    @Test
-    public void resultIsHandledCorrectlyWhenItShouldShowUI() throws IOException, URISyntaxException {
-        ImageComparison comparison = new ImageComparison("image1.png", "image2.png");
-        AtomicBoolean savedToFile = new AtomicBoolean(false);
-        AtomicBoolean showUI = new AtomicBoolean(false);
-
-        CommandLineUtil.handleResult(comparison, file -> savedToFile.set(true), () -> showUI.set(true));
-
-        assertFalse(savedToFile.get());
-        assertTrue(showUI.get());
-    }
-
-    @Test
-    public void resultIsHandledCorrectlyWhenItShouldSaveToFile() throws IOException, URISyntaxException {
-        File image1 = new File(ImageComparison.class.getClassLoader().getResource("image1.png").toURI().getPath());
-        File image2 = new File(ImageComparison.class.getClassLoader().getResource("image2.png").toURI().getPath());
-        File destination = Files.createTempFile("image-comparison-test", ".png").toFile();
-        ImageComparison comparison = CommandLineUtil.create(new ArgsParser.Arguments(image1, image2, destination));
-
-        AtomicBoolean savedToFile = new AtomicBoolean(false);
-        AtomicBoolean showUI = new AtomicBoolean(false);
-
-        CommandLineUtil.handleResult(comparison, file -> savedToFile.set(true), () -> showUI.set(true));
-
-        assertTrue(savedToFile.get());
-        assertFalse(showUI.get());
-    }
-
-    @Test
-    public void testSetterAndGettersThresholdWhenItShouldSetsProperly() throws IOException, URISyntaxException {
-        File image1 = new File(ImageComparison.class.getClassLoader().getResource("image1.png").toURI().getPath());
-        File image2 = new File(ImageComparison.class.getClassLoader().getResource("image2.png").toURI().getPath());
-        File destination = Files.createTempFile("image-comparison-test", ".png").toFile();
-        ImageComparison comparison = CommandLineUtil.create(new ArgsParser.Arguments(image1, image2, destination));
-        int setValue = 10;
-        comparison.setThreshold(setValue);
-        assertEquals(setValue, comparison.getThreshold());
-    }
-
-    @Test
     public void testGettersAnsSetters() throws IOException, URISyntaxException {
         //given
         ImageComparison imageComparison = new ImageComparison("image1.png", "image2.png");
@@ -272,22 +184,5 @@ public class ImageComparisonUnitTest {
         assertEquals(String.valueOf(200), String.valueOf(imageComparison.getMaximalRectangleCount()));
         assertEquals(String.valueOf(300), String.valueOf(imageComparison.getRectangleLineWidth()));
         assertEquals(String.valueOf(400), String.valueOf(imageComparison.getThreshold()));
-    }
-
-    private static void assertImagesEqual(BufferedImage imgA, BufferedImage imgB) {
-        if (imgA.getWidth() != imgB.getWidth() || imgA.getHeight() != imgB.getHeight()) {
-            fail("Images have different dimensions");
-        }
-
-        int width = imgA.getWidth();
-        int height = imgA.getHeight();
-
-        for (int y = 0; y < height; y++) {
-            for (int x = 0; x < width; x++) {
-                if (imgA.getRGB(x, y) != imgB.getRGB(x, y)) {
-                    fail("Images are different, found different pixel at: x = " + x + ", y = " + y);
-                }
-            }
-        }
     }
 }
