+++ src/test/java/com/github/romankh3/image/comparison/ImageComparisonUtilUnitTest.java
@@ -62,7 +62,7 @@ public class ImageComparisonUtilUnitTest {
     @Test
     public void testCommanlineUtilCreation() {
         //when
-        CommandLineUtil commandLineUtil = new CommandLineUtil();
+        CommandLineUsage commandLineUtil = new CommandLineUsage();
 
         //then
         assertNotNull(commandLineUtil);
